Use PRIuS/PRIdS macros to printf size_t/ssize_t. (#746)
PRIuS is defined to either "zu" or "Iu" depending on the platform
allowing to print size_t values in MSYS2. This macro is meant to be used
similarly to PRIu32, so that the "%" is not part of the macro allowing
expressions like `"%3" PRIuS` instead of `"%3zu"`. Similarly, `PRIdS` is
used for ssize_t.

This patch replaces all the occurrences and adds a test to bash_test.sh.

Fixes #718.

+++ examples/jxlinfo.c
@@ -5,6 +5,8 @@
 
 // This example prints information from the main codestream header.
 
+#include <inttypes.h>
+#include <stdint.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
@@ -232,7 +234,7 @@ int PrintBasicInfo(FILE* file) {
           fprintf(stderr, "JxlDecoderGetICCProfileSize failed\n");
           continue;
         }
-        printf("  ICC profile size: %zu\n", profile_size);
+        printf("  ICC profile size: %" PRIu64 "\n", (uint64_t)profile_size);
         if (profile_size < 132) {
           fprintf(stderr, "ICC profile too small\n");
           continue;
@@ -284,8 +286,8 @@ int PrintBasicInfo(FILE* file) {
       uint64_t size;
       JxlDecoderGetBoxType(dec, type, JXL_FALSE);
       JxlDecoderGetBoxSizeRaw(dec, &size);
-      printf("box: type: \"%c%c%c%c\" size: %zu\n", type[0], type[1], type[2],
-             type[3], (size_t)size);
+      printf("box: type: \"%c%c%c%c\" size: %" PRIu64 "\n", type[0], type[1],
+             type[2], type[3], (uint64_t)size);
     } else {
       fprintf(stderr, "Unexpected decoder status\n");
       break;
